+++ embed/config_logging_journal_windows.go
